Header and Footer components for landing page and its paths
