+++ app/component/mainHeader.js
@@ -1,0 +1,46 @@
+import GreenAppIcon from "@/public/greenAppIcon.svg";
+
+const navigation = [
+  { name: "Rewards", href: "/reward" },
+  { name: "Features", href: "/features" },
+  { name: "Marketplace", href: "#" },
+  { name: "Company", href: "#" },
+];
+
+export default function MainHeader() {
+  return (
+    <header className="absolute inset-x-0 top-0 z-50">
+      <nav
+        className="flex items-center justify-between p-6 lg:px-8"
+        aria-label="Global"
+      >
+        <div className="flex lg:flex-1">
+          <a href="#" className="-m-1.5 p-1.5">
+            <span className="sr-only">Your Company</span>
+            <GreenAppIcon className="h-16 w-32" />
+          </a>
+        </div>
+
+        <div className="hidden lg:flex lg:gap-x-12">
+          {navigation.map((item) => (
+            <a
+              key={item.name}
+              href={item.href}
+              className="text-sm font-semibold leading-6 text-gray-900"
+            >
+              {item.name}
+            </a>
+          ))}
+        </div>
+        <div className="hidden lg:flex lg:flex-1 lg:justify-end">
+          <a
+            href="/login"
+            className="text-sm font-semibold leading-6 text-gray-900"
+          >
+            Log in <span aria-hidden="true">&rarr;</span>
+          </a>
+        </div>
+      </nav>
+    </header>
+  );
+}
